remove the last of the things named gateway
