+++ router/handler_edge_ctrl/apiSessionUpdated.go
@@ -20,39 +20,34 @@ import (
 	"github.com/golang/protobuf/proto"
 	"github.com/michaelquigley/pfxlog"
 	"github.com/openziti/edge/controller/env"
-	"github.com/openziti/edge/gateway/internal/fabric"
+	"github.com/openziti/edge/router/internal/fabric"
 	"github.com/openziti/edge/pb/edge_ctrl_pb"
 	"github.com/openziti/foundation/channel2"
 )
 
-type sessionAddedHandler struct {
+type apiSessionUpdatedHandler struct {
 	sm fabric.StateManager
 }
 
-func NewSessionAddedHandler(sm fabric.StateManager) *sessionAddedHandler {
-	return &sessionAddedHandler{
+func NewApiSessionUpdatedHandler(sm fabric.StateManager) *apiSessionUpdatedHandler {
+	return &apiSessionUpdatedHandler{
 		sm: sm,
 	}
 }
 
-func (h *sessionAddedHandler) ContentType() int32 {
-	return env.SessionAddedType
+func (h *apiSessionUpdatedHandler) ContentType() int32 {
+	return env.ApiSessionUpdatedType
 }
 
-func (h *sessionAddedHandler) HandleReceive(msg *channel2.Message, _ channel2.Channel) {
+func (h *apiSessionUpdatedHandler) HandleReceive(msg *channel2.Message, ch channel2.Channel) {
 	go func() {
-		req := &edge_ctrl_pb.SessionAdded{}
+		req := &edge_ctrl_pb.ApiSessionUpdated{}
 		if err := proto.Unmarshal(msg.Body, req); err == nil {
-			for _, session := range req.Sessions {
-				pfxlog.Logger().Debugf("received new session %+v", session)
-				h.sm.AddSession(session)
-			}
-
-			if req.IsFullState {
-				h.sm.RemoveMissingSessions(req.Sessions)
+			for _, session := range req.ApiSessions {
+				h.sm.UpdateApiSession(session)
 			}
 		} else {
-			pfxlog.Logger().Panic("could not convert message as network session added")
+			pfxlog.Logger().Panic("could not convert message as network session updated")
 		}
 	}()
 }
