+++ router/handler_edge_ctrl/sessionRemoved.go
@@ -20,38 +20,34 @@ import (
 	"github.com/golang/protobuf/proto"
 	"github.com/michaelquigley/pfxlog"
 	"github.com/openziti/edge/controller/env"
-	"github.com/openziti/edge/gateway/internal/fabric"
+	"github.com/openziti/edge/router/internal/fabric"
 	"github.com/openziti/edge/pb/edge_ctrl_pb"
 	"github.com/openziti/foundation/channel2"
 )
 
-type apiSessionAddedHandler struct {
+type sessionRemovedHandler struct {
 	sm fabric.StateManager
 }
 
-func NewApiSessionAddedHandler(sm fabric.StateManager) *apiSessionAddedHandler {
-	return &apiSessionAddedHandler{
+func NewSessionRemovedHandler(sm fabric.StateManager) *sessionRemovedHandler {
+	return &sessionRemovedHandler{
 		sm: sm,
 	}
 }
 
-func (h *apiSessionAddedHandler) ContentType() int32 {
-	return env.ApiSessionAddedType
+func (h *sessionRemovedHandler) ContentType() int32 {
+	return env.SessionRemovedType
 }
 
-func (h *apiSessionAddedHandler) HandleReceive(msg *channel2.Message, ch channel2.Channel) {
+func (h *sessionRemovedHandler) HandleReceive(msg *channel2.Message, ch channel2.Channel) {
 	go func() {
-		req := &edge_ctrl_pb.ApiSessionAdded{}
+		req := &edge_ctrl_pb.SessionRemoved{}
 		if err := proto.Unmarshal(msg.Body, req); err == nil {
-			for _, session := range req.ApiSessions {
-				h.sm.AddApiSession(session)
-			}
-
-			if req.IsFullState {
-				h.sm.RemoveMissingApiSessions(req.ApiSessions)
+			for _, t := range req.Tokens {
+				h.sm.RemoveSession(t)
 			}
 		} else {
-			pfxlog.Logger().Panic("could not convert message as network session added")
+			pfxlog.Logger().Panic("could not convert message as network session removed")
 		}
 	}()
 }
