+++ router/handler_edge_ctrl/sessionUpdated.go
@@ -20,34 +20,34 @@ import (
 	"github.com/golang/protobuf/proto"
 	"github.com/michaelquigley/pfxlog"
 	"github.com/openziti/edge/controller/env"
-	"github.com/openziti/edge/gateway/internal/fabric"
+	"github.com/openziti/edge/router/internal/fabric"
 	"github.com/openziti/edge/pb/edge_ctrl_pb"
 	"github.com/openziti/foundation/channel2"
 )
 
-type apiSessionRemovedHandler struct {
+type sessionUpdatedHandler struct {
 	sm fabric.StateManager
 }
 
-func NewApiSessionRemovedHandler(sm fabric.StateManager) *apiSessionRemovedHandler {
-	return &apiSessionRemovedHandler{
+func NewSessionUpdatedHandler(sm fabric.StateManager) *sessionUpdatedHandler {
+	return &sessionUpdatedHandler{
 		sm: sm,
 	}
 }
 
-func (h *apiSessionRemovedHandler) ContentType() int32 {
-	return env.ApiSessionRemovedType
+func (h *sessionUpdatedHandler) ContentType() int32 {
+	return env.SessionUpdatedType
 }
 
-func (h *apiSessionRemovedHandler) HandleReceive(msg *channel2.Message, ch channel2.Channel) {
+func (h *sessionUpdatedHandler) HandleReceive(msg *channel2.Message, ch channel2.Channel) {
 	go func() {
-		req := &edge_ctrl_pb.ApiSessionRemoved{}
+		req := &edge_ctrl_pb.SessionAdded{}
 		if err := proto.Unmarshal(msg.Body, req); err == nil {
-			for _, t := range req.Tokens {
-				h.sm.RemoveApiSession(t)
+			for _, session := range req.Sessions {
+				h.sm.UpdateSession(session)
 			}
 		} else {
-			pfxlog.Logger().Panic("could not convert message as session removed")
+			pfxlog.Logger().Panic("could not convert message as network session updated")
 		}
 	}()
 }
