+++ router/internal/fabric/manager.go
