+++ router/internal/router/config.go
@@ -14,7 +14,7 @@
 	limitations under the License.
 */
 
-package gateway
+package router
 
 import (
 	"errors"
