+++ tests/context.go
@@ -29,8 +29,8 @@ import (
 	"encoding/pem"
 	"fmt"
 	"github.com/openziti/edge/eid"
-	"github.com/openziti/edge/gateway/enroll"
-	"github.com/openziti/edge/gateway/xgress_edge"
+	"github.com/openziti/edge/router/enroll"
+	"github.com/openziti/edge/router/xgress_edge"
 	"github.com/openziti/edge/rest_model"
 	"github.com/openziti/fabric/controller/xt_smartrouting"
 	"github.com/openziti/fabric/router"
